+++ MeepMeepTesting/src/main/java/com/example/meepmeeptesting/MeepMeepTesting.java
@@ -1,6 +1,9 @@
 package com.example.meepmeeptesting;
 
+import com.acmerobotics.roadrunner.Action;
 import com.acmerobotics.roadrunner.Pose2d;
+import com.acmerobotics.roadrunner.SequentialAction;
+import com.acmerobotics.roadrunner.TrajectoryActionBuilder;
 import com.acmerobotics.roadrunner.Vector2d;
 import com.noahbres.meepmeep.MeepMeep;
 import com.noahbres.meepmeep.roadrunner.DefaultBotBuilder;
@@ -17,10 +20,48 @@ public class MeepMeepTesting {
                 .setDriveTrainType(DriveTrainType.MECANUM)
                 .build();
 
-        myBot.runAction(myBot.getDrive().actionBuilder(new Pose2d(24, -64, 90))
-                        .splineTo(new Vector2d(30,30), Math.toRadians(0))
-                        .splineTo(new Vector2d(-20,-20), Math.toRadians(90))
-                .build());
+        Pose2d initialPose = new Pose2d(56, -52, Math.toRadians(90));
+
+        // vision here that outputs position
+        int visionOutputPosition = 1;
+
+        TrajectoryActionBuilder tab1Traj = myBot.getDrive().actionBuilder(initialPose)
+                .waitSeconds(1)
+                .lineToYSplineHeading(-36, Math.toRadians(0))
+                .setTangent(Math.toRadians(90))
+                .setTangent(Math.toRadians(0))
+                .lineToX(32)
+                .strafeTo(new Vector2d(44.5, 30))
+                .turn(Math.toRadians(180))
+                .lineToX(47.5);
+
+        Action tab1 = tab1Traj.build();
+
+        Action tab2 = myBot.getDrive().actionBuilder(initialPose)
+                .lineToY(37)
+                .setTangent(Math.toRadians(0))
+                .lineToX(18)
+                .waitSeconds(3)
+                .setTangent(Math.toRadians(0))
+                .lineToXSplineHeading(46, Math.toRadians(180))
+                .waitSeconds(3)
+                .build();
+
+        Action tab3 = myBot.getDrive().actionBuilder(initialPose)
+                .lineToYSplineHeading(33, Math.toRadians(180))
+                .waitSeconds(2)
+                .strafeTo(new Vector2d(46, 30))
+                .waitSeconds(3)
+                .build();
+
+        Action trajectoryActionCloseOut = tab1Traj.fresh()
+                .strafeTo(initialPose.component1()).build();
+
+        // actions that need to happen on init; for instance, a claw tightening.
+        myBot.runAction( new SequentialAction(
+                tab1,
+                trajectoryActionCloseOut));
+
 
         meepMeep.setBackground(MeepMeep.Background.FIELD_INTO_THE_DEEP_JUICE_DARK)
                 .setDarkMode(true)
