--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/BuiltInActions.java
@@ -1,4 +1,0 @@
-package org.firstinspires.ftc.teamcode;
-
-public class BuiltInActions {
-}
