+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/ExampleTeleOp.java
@@ -1,0 +1,17 @@
+package org.firstinspires.ftc.teamcode.opmodes;
+
+import com.qualcomm.robotcore.eventloop.opmode.*;
+
+@TeleOp(name = "ExampleTeleOp")
+public class ExampleTeleOp extends LinearOpMode {
+    @Override
+    public void runOpMode() {
+        waitForStart();
+        if (opModeIsActive()) {
+            // Pre-run
+            while (opModeIsActive()) {
+                // OpMode loop
+            }
+        }
+    }
+}
