+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/camera/CameraControl.java
@@ -1,0 +1,4 @@
+package org.firstinspires.ftc.teamcode.subsystems.camera;
+
+public class CameraControl {
+}
